fix: resolve did core service url
