+++ pkg/doc/vc/signer.go
@@ -30,5 +30,5 @@ type Signer struct {
 	Format                  vcsverifiable.Format               // VC format - LDP/JWT.
 	SignatureRepresentation verifiable.SignatureRepresentation // For LDP only.
 	KMS                     keyManager
-	VCStatusListVersion     StatusVersion // Version of VC status list
+	VCStatusListType        StatusType // Type of VC status list
 }
