+++ pkg/service/credentialstatus/statusProcessor.go
@@ -12,7 +12,7 @@ import (
 	"github.com/hyperledger/aries-framework-go/pkg/doc/verifiable"
 
 	"github.com/trustbloc/vcs/pkg/doc/vc"
-	"github.com/trustbloc/vcs/pkg/service/credentialstatus/versions"
+	"github.com/trustbloc/vcs/pkg/service/credentialstatus/statustype"
 )
 
 // VcStatusProcessor holds the list of methods required for processing different versions of Status(Revocation) List VC.
@@ -26,15 +26,15 @@ type VcStatusProcessor interface {
 }
 
 // GetVCStatusProcessor returns VcStatusProcessor.
-func GetVCStatusProcessor(vcStatusListVersion vc.StatusVersion) (VcStatusProcessor, error) {
-	switch vcStatusListVersion {
+func GetVCStatusProcessor(vcStatusListType vc.StatusType) (VcStatusProcessor, error) {
+	switch vcStatusListType {
 	case vc.StatusList2021VCStatus:
-		return versions.NewStatusList2021Processor(), nil
+		return statustype.NewStatusList2021Processor(), nil
 	case vc.RevocationList2021VCStatus:
-		return versions.NewRevocationList2021Processor(), nil
+		return statustype.NewRevocationList2021Processor(), nil
 	case vc.RevocationList2020VCStatus:
-		return versions.NewRevocationList2020Processor(), nil
+		return statustype.NewRevocationList2020Processor(), nil
 	default:
-		return nil, fmt.Errorf("unsupported VCStatusListVersion %s", vcStatusListVersion)
+		return nil, fmt.Errorf("unsupported VCStatusListType %s", vcStatusListType)
 	}
 }
