+++ pkg/service/credentialstatus/statustype/common.go
@@ -4,7 +4,7 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package versions
+package statustype
 
 type credentialSubject struct {
 	ID            string `json:"id"`
