+++ pkg/service/credentialstatus/statustype/revocationlist2020.go
@@ -4,7 +4,7 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package versions
+package statustype
 
 import (
 	"fmt"
