+++ pkg/service/verifycredential/revocation/identityhub_test.go
@@ -30,10 +30,10 @@ import (
 )
 
 const (
-	objectID                 = "33c34640-8e48-43ba-9566-32760ab2cdad"
-	didID                    = "did:ion:EiCMdVLtzqqW5n6zUC3_srZxWPCseVxKXu9FqQ8LyS1mTA:eyJkZWx0YSI6eyJwYXRjaGVzIjpbeyJhY3Rpb24iOiJyZXBsYWNlIiwiZG9jdW1lbnQiOnsicHVibGljS2V5cyI6W3siaWQiOiI2NmRkNTFmZTBjYWM0ZjFhYWU4MTJkMGFhMTA5YmMyYXZjU2lnbmluZ0tleS0yZTk3NSIsInB1YmxpY0tleUp3ayI6eyJjcnYiOiJzZWNwMjU2azEiLCJrdHkiOiJFQyIsIngiOiJqNVQ4S1FfQ19IRGxSbXlFX1pwRjltbE1RZ3B4N19fMFJQRHhPVmM4dWt3IiwieSI6InpybDBWSllHWnhVLXFjZWt2SlY4NGs5U2x2STQxam53NG4yTS1WMnB4MGMifSwicHVycG9zZXMiOlsiYXV0aGVudGljYXRpb24iLCJhc3NlcnRpb25NZXRob2QiXSwidHlwZSI6IkVjZHNhU2VjcDI1NmsxVmVyaWZpY2F0aW9uS2V5MjAxOSJ9XSwic2VydmljZXMiOlt7ImlkIjoibGlua2VkZG9tYWlucyIsInNlcnZpY2VFbmRwb2ludCI6eyJvcmlnaW5zIjpbImh0dHBzOi8vZGlkLnJvaGl0Z3VsYXRpLmNvbS8iXX0sInR5cGUiOiJMaW5rZWREb21haW5zIn0seyJpZCI6Imh1YiIsInNlcnZpY2VFbmRwb2ludCI6eyJpbnN0YW5jZXMiOlsiaHR0cHM6Ly9iZXRhLmh1Yi5tc2lkZW50aXR5LmNvbS92MS4wL2E0OTJjZmYyLWQ3MzMtNDA1Ny05NWE1LWE3MWZjMzY5NWJjOCJdfSwidHlwZSI6IklkZW50aXR5SHViIn1dfX1dLCJ1cGRhdGVDb21taXRtZW50IjoiRWlDcXRpZnUwSHg4RUVkbGlrVnZIWGpYZzRLb0pZZUV0cDdZeGlvRzVYWmRKZyJ9LCJzdWZmaXhEYXRhIjp7ImRlbHRhSGFzaCI6IkVpQ1NVQklmYTBXZHBXNm5oVTdNaHlSczRucTFDeEg1V1ZyUjVkUFZYV09MYmciLCJyZWNvdmVyeUNvbW1pdG1lbnQiOiJFaUF1cGoxRWZsOHdjWlRQZTI3X0lGWEJ3MjlzOEN5SXBRX3UzVkRwUmswdkNRIn19" //nolint:lll
-	malformedQueriesURLParam = "W3sic2NoZW1hIjoiaHR0cHM6Ly93M2lkLm9yZy92Yy1zdGF0dXMtbGlzdC0yMDIxL3YxIiwib2JqZWN0SWQiOiIzM2MzNDY0MC04ZTQ4LTQzYmEtOTU2Ni0zMjc2MGFiMmNkYWQifV0="                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        //nolint:lll
-	queriesURLParam          = "W3sibWV0aG9kIjoiQ29sbGVjdGlvbnNRdWVyeSIsInNjaGVtYSI6Imh0dHBzOi8vdzNpZC5vcmcvdmMtc3RhdHVzLWxpc3QtMjAyMS92MSIsIm9iamVjdElkIjoiMzNjMzQ2NDAtOGU0OC00M2JhLTk1NjYtMzI3NjBhYjJjZGFkIn1d"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    //nolint:lll
+	objectID                 = "686b6953-a0a4-4ca8-b9d9-b8756e1aa3dd"
+	didID                    = "did:ion:EiCB7y_BnrO1nfsfpqlUANEcW8QpcjHWxK3mjiOSbD9ptQ:eyJkZWx0YSI6eyJwYXRjaGVzIjpbeyJhY3Rpb24iOiJyZXBsYWNlIiwiZG9jdW1lbnQiOnsicHVibGljS2V5cyI6W3siaWQiOiI3NzAxMjE2NTAzYjU0MjRjYmNlY2RjN2EyZjQwZDkzOHZjU2lnbmluZ0tleS04NDYxNyIsInB1YmxpY0tleUp3ayI6eyJjcnYiOiJzZWNwMjU2azEiLCJrdHkiOiJFQyIsIngiOiIwaHZfVnZkS1FPd04xT2l4Q0VoOE5Gd2RHdUJMZjZRYWJUenZxZ2VfNWpzIiwieSI6Il8wdTVGUGthTUtKSm5heGU4ZFAxbGNUakJEQ1RHdkdyR21iLU9waDdvY2cifSwicHVycG9zZXMiOlsiYXV0aGVudGljYXRpb24iLCJhc3NlcnRpb25NZXRob2QiXSwidHlwZSI6IkVjZHNhU2VjcDI1NmsxVmVyaWZpY2F0aW9uS2V5MjAxOSJ9XSwic2VydmljZXMiOlt7ImlkIjoibGlua2VkZG9tYWlucyIsInNlcnZpY2VFbmRwb2ludCI6eyJvcmlnaW5zIjpbImh0dHBzOi8vZGlkLnJvaGl0Z3VsYXRpLmNvbS8iXX0sInR5cGUiOiJMaW5rZWREb21haW5zIn0seyJpZCI6Imh1YiIsInNlcnZpY2VFbmRwb2ludCI6eyJpbnN0YW5jZXMiOlsiaHR0cHM6Ly9odWIuZGlkLm1zaWRlbnRpdHkuY29tL3YxLjAvYTQ5MmNmZjItZDczMy00MDU3LTk1YTUtYTcxZmMzNjk1YmM4Il19LCJ0eXBlIjoiSWRlbnRpdHlIdWIifV19fV0sInVwZGF0ZUNvbW1pdG1lbnQiOiJFaUNmNUhGZjR5S0lUcWRzTHpQbDhjcHRObTR5Y1g3dXZjcnBFMW5wWnZ2Qmt3In0sInN1ZmZpeERhdGEiOnsiZGVsdGFIYXNoIjoiRWlETjEyZ1BXcGNfUXJaYk1NSUNpQVJ5aWFqbGRKLXVWTjlqZWRTRWtBbWVzdyIsInJlY292ZXJ5Q29tbWl0bWVudCI6IkVpQWNJQjFTeW83eXE3bDdkYWJ2T2lTdmdnZmpfcGR0SlV2cUgwb05UNVNDSWcifX0" //nolint:lll
+	malformedQueriesURLParam = "W3sic2NoZW1hIjoiaHR0cHM6Ly93M2lkLm9yZy92Yy1zdGF0dXMtbGlzdC0yMDIxL3YxIiwib2JqZWN0SWQiOiIzM2MzNDY0MC04ZTQ4LTQzYmEtOTU2Ni0zMjc2MGFiMmNkYWQifV0="                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       //nolint:lll
+	queriesURLParam          = "W3sibWV0aG9kIjoiQ29sbGVjdGlvbnNRdWVyeSIsInNjaGVtYSI6Imh0dHBzOi8vdzNpZC5vcmcvdmMtc3RhdHVzLWxpc3QtMjAyMS92MSIsIm9iamVjdElkIjoiNjg2YjY5NTMtYTBhNC00Y2E4LWI5ZDktYjg3NTZlMWFhM2RkIn1d"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   //nolint:lll
 	query                    = "service=IdentityHub&queries=" + queriesURLParam
 	didRelativeURL           = didID + "?" + query
 	serviceEndpointURL       = "https://identityhub.example.com/"
@@ -266,6 +266,26 @@ func TestService_resolveDIDRelativeUrl(t *testing.T) {
 						},
 						doErr: nil,
 					}
+				},
+			},
+			args: args{
+				didRelativeURL: didRelativeURL,
+			},
+			want:    sampleVCJWT,
+			wantErr: false,
+		},
+		{
+			name: "OK - uniresolver",
+			fields: fields{
+				getVDR: func() vdr.Registry {
+					universalResolverVDRI, err := httpbinding.New("https://uniresolver.io/1.0/identifiers",
+						httpbinding.WithAccept(func(method string) bool { return method == "ion" }))
+					require.NoError(t, err)
+
+					return vdr2.New(vdr2.WithVDR(universalResolverVDRI))
+				},
+				getHTTPClient: func() httpClient {
+					return http.DefaultClient
 				},
 			},
 			args: args{
