--- pkg/service/verifycredential/revocation/testdata/identity_hub_response_ldp.json
@@ -1,27 +1,0 @@
-{
-  "requestId": "b74373c9-ace3-48c6-81e2-b8436b3d642e",
-  "status": {
-    "code": 200
-  },
-  "replies": [
-    {
-      "messageId": "bafkreicug4tkpif4c4f2ud3x7dwyte7pcc74667uz4x3jzth2espi5f7g4",
-      "status": {
-        "code": 200
-      },
-      "entries": [
-        {
-          "descriptor": {
-            "method": "CollectionsWrite",
-            "schema": "https://w3id.org/vc-status-list-2021/v1",
-            "dataFormat": "application/vc+ldp",
-            "objectId": "33c34640-8e48-43ba-9566-32760ab2cdad",
-            "clock": 0,
-            "cid": "bafkreifen7lwem5744ht7korrvqr5p3wxjtckzkpccliunnblsdcupoxve"
-          },
-          "data": "ewogICJAY29udGV4dCI6IFsKICAgICJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSIsCiAgICAiaHR0cHM6Ly93M2MtY2NnLmdpdGh1Yi5pby9sZHMtandzMjAyMC9jb250ZXh0cy9sZHMtandzMjAyMC12MS5qc29uIiwKICAgICJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy9leGFtcGxlcy92MSIsCiAgICAiaHR0cHM6Ly93M2lkLm9yZy92Yy9zdGF0dXMtbGlzdC8yMDIxL3YxIgogIF0sCiAgInR5cGUiOiBbCiAgICAiVmVyaWZpYWJsZUNyZWRlbnRpYWwiLAogICAgIlVuaXZlcnNpdHlEZWdyZWVDcmVkZW50aWFsIgogIF0sCiAgImlkIjogImh0dHA6Ly9leGFtcGxlLmdvdi9jcmVkZW50aWFscy8zNzMyIiwKICAiaXNzdWFuY2VEYXRlIjogIjIwMjAtMDMtMTZUMjI6Mzc6MjYuNTQ0WiIsCiAgImlzc3VlciI6IHsKICAgICJpZCI6ICJkaWQ6dHJ1c3RibG9jazphYmMiLAogICAgIm5hbWUiOiAiVW5pdmVyc2l0eSIKICB9LAogICJjcmVkZW50aWFsU3RhdHVzIjogewogICAgImlkIjogImh0dHBzOi8vaXNzdWVyLXZjcy5zYW5kYm94LnRydXN0YmxvYy5kZXYvdmMtaXNzdWVyLXRlc3QtMi9zdGF0dXMvMSMwIiwKICAgICJ0eXBlIjogIlN0YXR1c0xpc3QyMDIxRW50cnkiLAogICAgInN0YXR1c0xpc3RJbmRleCI6ICIxIiwKICAgICJzdGF0dXNMaXN0Q3JlZGVudGlhbCI6ICIiLAogICAgInN0YXR1c1B1cnBvc2UiOiAiMiIKICB9LAogICJjcmVkZW50aWFsU3ViamVjdCI6IHsKICAgICJpZCI6ICJkaWQ6ZXhhbXBsZTplYmZlYjFmNzEyZWJjNmYxYzI3NmUxMmVjMjEiLAogICAgImRlZ3JlZSI6IHsKICAgICAgInR5cGUiOiAiQmFjaGVsb3JEZWdyZWUiLAogICAgICAiZGVncmVlIjogIk1JVCIKICAgIH0sCiAgICAibmFtZSI6ICJKYXlkZW4gRG9lIiwKICAgICJzcG91c2UiOiAiZGlkOmV4YW1wbGU6YzI3NmUxMmVjMjFlYmZlYjFmNzEyZWJjNmYxIgogIH0KfQ=="
-        }
-      ]
-    }
-  ]
-}
