+++ pkg/service/verifycredential/revocation/testdata/sample_vc.jwt
@@ -1,1 +1,1 @@
-eyJhbGciOiJub25lIiwidHlwIjoiSldUIn0.eyJleHAiOjE1Nzc5MDY2MDQsImlhdCI6MTI2MjM3MzgwNCwiaXNzIjoiZGlkOmV4YW1wbGU6NzZlMTJlYzcxMmViYzZmMWMyMjFlYmZlYjFmIiwianRpIjoiaHR0cDovL2V4YW1wbGUuZWR1L2NyZWRlbnRpYWxzLzE4NzIiLCJuYmYiOjEyNjIzNzM4MDQsInN1YiI6ImRpZDpleGFtcGxlOmViZmViMWY3MTJlYmM2ZjFjMjc2ZTEyZWMyMSIsInZjIjp7IkBjb250ZXh0IjpbImh0dHBzOi8vd3d3LnczLm9yZy8yMDE4L2NyZWRlbnRpYWxzL3YxIiwiaHR0cHM6Ly93d3cudzMub3JnLzIwMTgvY3JlZGVudGlhbHMvZXhhbXBsZXMvdjEiLCJodHRwczovL3czaWQub3JnL3NlY3VyaXR5L2Jicy92MSJdLCJjcmVkZW50aWFsU3ViamVjdCI6eyJkZWdyZWUiOnsidHlwZSI6IkJhY2hlbG9yRGVncmVlIiwidW5pdmVyc2l0eSI6Ik1JVCJ9LCJpZCI6ImRpZDpleGFtcGxlOmViZmViMWY3MTJlYmM2ZjFjMjc2ZTEyZWMyMSIsIm5hbWUiOiJKYXlkZW4gRG9lIiwic3BvdXNlIjoiZGlkOmV4YW1wbGU6YzI3NmUxMmVjMjFlYmZlYjFmNzEyZWJjNmYxIn0sImV4cGlyYXRpb25EYXRlIjoiMjAyMC0wMS0wMVQxOToyMzoyNFoiLCJpZCI6Imh0dHA6Ly9leGFtcGxlLmVkdS9jcmVkZW50aWFscy8xODcyIiwiaXNzdWFuY2VEYXRlIjoiMjAxMC0wMS0wMVQxOToyMzoyNFoiLCJpc3N1ZXIiOnsiaWQiOiJkaWQ6ZXhhbXBsZTo3NmUxMmVjNzEyZWJjNmYxYzIyMWViZmViMWYiLCJuYW1lIjoiRXhhbXBsZSBVbml2ZXJzaXR5In0sInJlZmVyZW5jZU51bWJlciI6OC4zMjk0ODQ3ZSswNywidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsIlVuaXZlcnNpdHlEZWdyZWVDcmVkZW50aWFsIl19fQ.
+eyJhbGciOiJFUzI1NksiLCJraWQiOiJkaWQ6aW9uOkVpQ0I3eV9CbnJPMW5mc2ZwcWxVQU5FY1c4UXBjakhXeEszbWppT1NiRDlwdFE6ZXlKa1pXeDBZU0k2ZXlKd1lYUmphR1Z6SWpwYmV5SmhZM1JwYjI0aU9pSnlaWEJzWVdObElpd2laRzlqZFcxbGJuUWlPbnNpY0hWaWJHbGpTMlY1Y3lJNlczc2lhV1FpT2lJM056QXhNakUyTlRBellqVTBNalJqWW1ObFkyUmpOMkV5WmpRd1pEa3pPSFpqVTJsbmJtbHVaMHRsZVMwNE5EWXhOeUlzSW5CMVlteHBZMHRsZVVwM2F5STZleUpqY25ZaU9pSnpaV053TWpVMmF6RWlMQ0pyZEhraU9pSkZReUlzSW5naU9pSXdhSFpmVm5aa1MxRlBkMDR4VDJsNFEwVm9PRTVHZDJSSGRVSk1aalpSWVdKVWVuWnhaMlZmTldweklpd2llU0k2SWw4d2RUVkdVR3RoVFV0S1NtNWhlR1U0WkZBeGJHTlVha0pFUTFSSGRrZHlSMjFpTFU5d2FEZHZZMmNpZlN3aWNIVnljRzl6WlhNaU9sc2lZWFYwYUdWdWRHbGpZWFJwYjI0aUxDSmhjM05sY25ScGIyNU5aWFJvYjJRaVhTd2lkSGx3WlNJNklrVmpaSE5oVTJWamNESTFObXN4Vm1WeWFXWnBZMkYwYVc5dVMyVjVNakF4T1NKOVhTd2ljMlZ5ZG1salpYTWlPbHQ3SW1sa0lqb2liR2x1YTJWa1pHOXRZV2x1Y3lJc0luTmxjblpwWTJWRmJtUndiMmx1ZENJNmV5SnZjbWxuYVc1eklqcGJJbWgwZEhCek9pOHZaR2xrTG5KdmFHbDBaM1ZzWVhScExtTnZiUzhpWFgwc0luUjVjR1VpT2lKTWFXNXJaV1JFYjIxaGFXNXpJbjBzZXlKcFpDSTZJbWgxWWlJc0luTmxjblpwWTJWRmJtUndiMmx1ZENJNmV5SnBibk4wWVc1alpYTWlPbHNpYUhSMGNITTZMeTlvZFdJdVpHbGtMbTF6YVdSbGJuUnBkSGt1WTI5dEwzWXhMakF2WVRRNU1tTm1aakl0WkRjek15MDBNRFUzTFRrMVlUVXRZVGN4Wm1Nek5qazFZbU00SWwxOUxDSjBlWEJsSWpvaVNXUmxiblJwZEhsSWRXSWlmVjE5ZlYwc0luVndaR0YwWlVOdmJXMXBkRzFsYm5RaU9pSkZhVU5tTlVoR1pqUjVTMGxVY1dSelRIcFFiRGhqY0hST2JUUjVZMWczZFhaamNuQkZNVzV3V25aMlFtdDNJbjBzSW5OMVptWnBlRVJoZEdFaU9uc2laR1ZzZEdGSVlYTm9Jam9pUldsRVRqRXlaMUJYY0dOZlVYSmFZazFOU1VOcFFWSjVhV0ZxYkdSS0xYVldUamxxWldSVFJXdEJiV1Z6ZHlJc0luSmxZMjkyWlhKNVEyOXRiV2wwYldWdWRDSTZJa1ZwUVdOSlFqRlRlVzgzZVhFM2JEZGtZV0oyVDJsVGRtZG5abXBmY0dSMFNsVjJjVWd3YjA1VU5WTkRTV2NpZlgwIzc3MDEyMTY1MDNiNTQyNGNiY2VjZGM3YTJmNDBkOTM4dmNTaWduaW5nS2V5LTg0NjE3IiwidHlwIjoiSldUIn0.eyJ2YyI6eyJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSIsImh0dHBzOi8vdzNpZC5vcmcvdmMtc3RhdHVzLWxpc3QtMjAyMS92MSJdLCJ0eXBlIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsIiwiU3RhdHVzTGlzdDIwMjFDcmVkZW50aWFsIl0sImNyZWRlbnRpYWxTdWJqZWN0Ijp7ImlkIjoidXJuOnV1aWQ6Njg2YjY5NTMtYTBhNC00Y2E4LWI5ZDktYjg3NTZlMWFhM2RkIiwidHlwZSI6IlJldm9jYXRpb25MaXN0MjAyMSIsImVuY29kZWRMaXN0IjoiSDRzSUFBQUFBQUFBQS0zQkFRRUFBQUNDSVAtdmJraEFBUUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQWZCaEpsMkFWcUdFQUFBIn19LCJqdGkiOiJkaWQ6aW9uOkVpQ0I3eV9CbnJPMW5mc2ZwcWxVQU5FY1c4UXBjakhXeEszbWppT1NiRDlwdFE6ZXlKa1pXeDBZU0k2ZXlKd1lYUmphR1Z6SWpwYmV5SmhZM1JwYjI0aU9pSnlaWEJzWVdObElpd2laRzlqZFcxbGJuUWlPbnNpY0hWaWJHbGpTMlY1Y3lJNlczc2lhV1FpT2lJM056QXhNakUyTlRBellqVTBNalJqWW1ObFkyUmpOMkV5WmpRd1pEa3pPSFpqVTJsbmJtbHVaMHRsZVMwNE5EWXhOeUlzSW5CMVlteHBZMHRsZVVwM2F5STZleUpqY25ZaU9pSnpaV053TWpVMmF6RWlMQ0pyZEhraU9pSkZReUlzSW5naU9pSXdhSFpmVm5aa1MxRlBkMDR4VDJsNFEwVm9PRTVHZDJSSGRVSk1aalpSWVdKVWVuWnhaMlZmTldweklpd2llU0k2SWw4d2RUVkdVR3RoVFV0S1NtNWhlR1U0WkZBeGJHTlVha0pFUTFSSGRrZHlSMjFpTFU5d2FEZHZZMmNpZlN3aWNIVnljRzl6WlhNaU9sc2lZWFYwYUdWdWRHbGpZWFJwYjI0aUxDSmhjM05sY25ScGIyNU5aWFJvYjJRaVhTd2lkSGx3WlNJNklrVmpaSE5oVTJWamNESTFObXN4Vm1WeWFXWnBZMkYwYVc5dVMyVjVNakF4T1NKOVhTd2ljMlZ5ZG1salpYTWlPbHQ3SW1sa0lqb2liR2x1YTJWa1pHOXRZV2x1Y3lJc0luTmxjblpwWTJWRmJtUndiMmx1ZENJNmV5SnZjbWxuYVc1eklqcGJJbWgwZEhCek9pOHZaR2xrTG5KdmFHbDBaM1ZzWVhScExtTnZiUzhpWFgwc0luUjVjR1VpT2lKTWFXNXJaV1JFYjIxaGFXNXpJbjBzZXlKcFpDSTZJbWgxWWlJc0luTmxjblpwWTJWRmJtUndiMmx1ZENJNmV5SnBibk4wWVc1alpYTWlPbHNpYUhSMGNITTZMeTlvZFdJdVpHbGtMbTF6YVdSbGJuUnBkSGt1WTI5dEwzWXhMakF2WVRRNU1tTm1aakl0WkRjek15MDBNRFUzTFRrMVlUVXRZVGN4Wm1Nek5qazFZbU00SWwxOUxDSjBlWEJsSWpvaVNXUmxiblJwZEhsSWRXSWlmVjE5ZlYwc0luVndaR0YwWlVOdmJXMXBkRzFsYm5RaU9pSkZhVU5tTlVoR1pqUjVTMGxVY1dSelRIcFFiRGhqY0hST2JUUjVZMWczZFhaamNuQkZNVzV3V25aMlFtdDNJbjBzSW5OMVptWnBlRVJoZEdFaU9uc2laR1ZzZEdGSVlYTm9Jam9pUldsRVRqRXlaMUJYY0dOZlVYSmFZazFOU1VOcFFWSjVhV0ZxYkdSS0xYVldUamxxWldSVFJXdEJiV1Z6ZHlJc0luSmxZMjkyWlhKNVEyOXRiV2wwYldWdWRDSTZJa1ZwUVdOSlFqRlRlVzgzZVhFM2JEZGtZV0oyVDJsVGRtZG5abXBmY0dSMFNsVjJjVWd3YjA1VU5WTkRTV2NpZlgwP3NlcnZpY2U9SWRlbnRpdHlIdWImcXVlcmllcz1XM3NpYldWMGFHOWtJam9pUTI5c2JHVmpkR2x2Ym5OUmRXVnllU0lzSW5OamFHVnRZU0k2SW1oMGRIQnpPaTh2ZHpOcFpDNXZjbWN2ZG1NdGMzUmhkSFZ6TFd4cGMzUXRNakF5TVM5Mk1TSXNJbTlpYW1WamRFbGtJam9pTmpnMllqWTVOVE10WVRCaE5DMDBZMkU0TFdJNVpEa3RZamczTlRabE1XRmhNMlJrSW4xZCIsImlzcyI6ImRpZDppb246RWlDQjd5X0Juck8xbmZzZnBxbFVBTkVjVzhRcGNqSFd4SzNtamlPU2JEOXB0UTpleUprWld4MFlTSTZleUp3WVhSamFHVnpJanBiZXlKaFkzUnBiMjRpT2lKeVpYQnNZV05sSWl3aVpHOWpkVzFsYm5RaU9uc2ljSFZpYkdsalMyVjVjeUk2VzNzaWFXUWlPaUkzTnpBeE1qRTJOVEF6WWpVME1qUmpZbU5sWTJSak4yRXlaalF3WkRrek9IWmpVMmxuYm1sdVowdGxlUzA0TkRZeE55SXNJbkIxWW14cFkwdGxlVXAzYXlJNmV5SmpjbllpT2lKelpXTndNalUyYXpFaUxDSnJkSGtpT2lKRlF5SXNJbmdpT2lJd2FIWmZWblprUzFGUGQwNHhUMmw0UTBWb09FNUdkMlJIZFVKTVpqWlJZV0pVZW5aeFoyVmZOV3B6SWl3aWVTSTZJbDh3ZFRWR1VHdGhUVXRLU201aGVHVTRaRkF4YkdOVWFrSkVRMVJIZGtkeVIyMWlMVTl3YURkdlkyY2lmU3dpY0hWeWNHOXpaWE1pT2xzaVlYVjBhR1Z1ZEdsallYUnBiMjRpTENKaGMzTmxjblJwYjI1TlpYUm9iMlFpWFN3aWRIbHdaU0k2SWtWalpITmhVMlZqY0RJMU5tc3hWbVZ5YVdacFkyRjBhVzl1UzJWNU1qQXhPU0o5WFN3aWMyVnlkbWxqWlhNaU9sdDdJbWxrSWpvaWJHbHVhMlZrWkc5dFlXbHVjeUlzSW5ObGNuWnBZMlZGYm1Sd2IybHVkQ0k2ZXlKdmNtbG5hVzV6SWpwYkltaDBkSEJ6T2k4dlpHbGtMbkp2YUdsMFozVnNZWFJwTG1OdmJTOGlYWDBzSW5SNWNHVWlPaUpNYVc1clpXUkViMjFoYVc1ekluMHNleUpwWkNJNkltaDFZaUlzSW5ObGNuWnBZMlZGYm1Sd2IybHVkQ0k2ZXlKcGJuTjBZVzVqWlhNaU9sc2lhSFIwY0hNNkx5OW9kV0l1Wkdsa0xtMXphV1JsYm5ScGRIa3VZMjl0TDNZeExqQXZZVFE1TW1ObVpqSXRaRGN6TXkwME1EVTNMVGsxWVRVdFlUY3habU16TmprMVltTTRJbDE5TENKMGVYQmxJam9pU1dSbGJuUnBkSGxJZFdJaWZWMTlmVjBzSW5Wd1pHRjBaVU52YlcxcGRHMWxiblFpT2lKRmFVTm1OVWhHWmpSNVMwbFVjV1J6VEhwUWJEaGpjSFJPYlRSNVkxZzNkWFpqY25CRk1XNXdXbloyUW10M0luMHNJbk4xWm1acGVFUmhkR0VpT25zaVpHVnNkR0ZJWVhOb0lqb2lSV2xFVGpFeVoxQlhjR05mVVhKYVlrMU5TVU5wUVZKNWFXRnFiR1JLTFhWV1RqbHFaV1JUUld0QmJXVnpkeUlzSW5KbFkyOTJaWEo1UTI5dGJXbDBiV1Z1ZENJNklrVnBRV05KUWpGVGVXODNlWEUzYkRka1lXSjJUMmxUZG1kblptcGZjR1IwU2xWMmNVZ3diMDVVTlZORFNXY2lmWDAiLCJzdWIiOiJ1cm46dXVpZDo2ODZiNjk1My1hMGE0LTRjYTgtYjlkOS1iODc1NmUxYWEzZGQiLCJpYXQiOjE2NjczMzY3NDJ9.YLwCdupj3wGFJITBj3EvpYQpCFCfwaoitIdopEO3zaVSQbJaXJbbaMjHfB316BQAhuzLFrdOATiSRFi_Qu-1DQ
