+++ test/bdd/fixtures/profile/profiles.json
@@ -13,7 +13,9 @@
           "keyType": "ECDSASecp256k1IEEEP1363",
           "format": "ldp",
           "didMethod": "web",
-          "vcStatusListVersion": "StatusList2021Entry"
+          "status": {
+            "type": "StatusList2021Entry"
+          }
         }
       },
       "createDID": true,
@@ -33,7 +35,9 @@
           "keyType": "ECDSAP256DER",
           "format": "jwt",
           "didMethod": "orb",
-          "vcStatusListVersion": "StatusList2021Entry"
+          "status": {
+            "type": "StatusList2021Entry"
+          }
         }
       },
       "createDID": true,
@@ -53,7 +57,9 @@
           "keyType": "ECDSAP384DER",
           "format": "jwt",
           "didMethod": "orb",
-          "vcStatusListVersion": "RevocationList2021Status"
+          "status": {
+            "type": "RevocationList2021Status"
+          }
         }
       },
       "createDID": true,
@@ -73,7 +79,9 @@
           "keyType": "ECDSASecp256k1IEEEP1363",
           "format": "jwt",
           "didMethod": "orb",
-          "vcStatusListVersion": "RevocationList2020Status"
+          "status": {
+            "type": "RevocationList2020Status"
+          }
         }
       },
       "createDID": true,
@@ -93,7 +101,9 @@
           "keyType": "ECDSAP256DER",
           "format": "ldp",
           "didMethod": "web",
-          "vcStatusListVersion": "RevocationList2020Status"
+          "status": {
+            "type": "RevocationList2020Status"
+          }
         }
       },
       "createDID": true,
@@ -113,7 +123,9 @@
           "keyType": "ECDSASecp256k1IEEEP1363",
           "format": "ldp",
           "didMethod": "orb",
-          "vcStatusListVersion": "StatusList2021Entry"
+          "status": {
+            "type": "StatusList2021Entry"
+          }
         }
       },
       "createDID": true,
@@ -133,7 +145,9 @@
           "keyType": "ECDSAP256DER",
           "format": "ldp",
           "didMethod": "orb",
-          "vcStatusListVersion": "RevocationList2020Status"
+          "status": {
+            "type": "RevocationList2020Status"
+          }
         }
       },
       "createDID": true,
@@ -153,7 +167,9 @@
           "keyType": "ECDSAP384DER",
           "format": "ldp",
           "didMethod": "orb",
-          "vcStatusListVersion": "StatusList2021Entry"
+          "status": {
+            "type": "StatusList2021Entry"
+          }
         }
       },
       "createDID": true,
@@ -173,7 +189,9 @@
           "keyType": "ECDSAP256DER",
           "format": "ldp",
           "didMethod": "orb",
-          "vcStatusListVersion": "RevocationList2020Status"
+          "status": {
+            "type": "RevocationList2020Status"
+          }
         }
       },
       "createDID": true,
@@ -193,7 +211,9 @@
           "keyType": "ECDSAP384DER",
           "format": "ldp",
           "didMethod": "key",
-          "vcStatusListVersion": "StatusList2021Entry"
+          "status": {
+            "type": "StatusList2021Entry"
+          }
         }
       },
       "createDID": true
@@ -211,7 +231,9 @@
           "keyType": "ECDSAP256DER",
           "format": "ldp",
           "didMethod": "key",
-          "vcStatusListVersion": "RevocationList2020Status"
+          "status": {
+            "type": "RevocationList2020Status"
+          }
         }
       },
       "createDID": true
@@ -229,7 +251,9 @@
           "keyType": "ECDSASecp256k1IEEEP1363",
           "format": "ldp",
           "didMethod": "orb",
-          "vcStatusListVersion": "StatusList2021Entry"
+          "status": {
+            "type": "StatusList2021Entry"
+          }
         },
         "oidcConfig": {
           "client_id": "bank_issuer",
